+++ sites/nau/src/backend/nau/force_translations.py
@@ -5,3 +5,21 @@ from django.utils.translation import gettext_lazy as _
 _("{!s} settings")
 _("First session")
 _("Persons")
+
+_("Sub categories")
+_("Related courses")
+_("Courses pagination")
+_("See all courses")
+_("Related organizations")
+_("Related organizations pagination")
+_("Related blogposts")
+_("Related blogposts pagination")
+_("Related persons")
+_("Related persons pagination")
+_("No other open course runs")
+_("No open course runs")
+_("Life-changing learning!")
+
+# Richie apps/core/templates/social-networks/course-badges.html
+_("Follow a course online with Richie")
+_('I just enrolled to the course "%(title)s" on Richie: %(url)s')